feat: new tests, small changes, test configurations and final touches
