+++ public/test-results/test-output.json
@@ -1,0 +1,1 @@
+{"numTotalTestSuites":2,"numPassedTestSuites":2,"numFailedTestSuites":0,"numPendingTestSuites":0,"numTotalTests":10,"numPassedTests":10,"numFailedTests":0,"numPendingTests":0,"numTodoTests":0,"startTime":1723061932085,"success":true,"testResults":[{"assertionResults":[{"ancestorTitles":[""],"fullName":" creates a count function","status":"passed","title":"creates a count function","duration":8,"failureMessages":[]},{"ancestorTitles":[""],"fullName":" creates a count starting from 10 with a step of 5","status":"passed","title":"creates a count starting from 10 with a step of 5","duration":1,"failureMessages":[]},{"ancestorTitles":[""],"fullName":" defaults to start 0, step 1 when no arguments passed","status":"passed","title":"defaults to start 0, step 1 when no arguments passed","duration":1,"failureMessages":[]}],"startTime":1723061932651,"endTime":1723061932661,"status":"passed","message":"","name":"/Users/felipesilva/Flip/centered_counter/src/factory.test.ts"},{"assertionResults":[{"ancestorTitles":[""],"fullName":" creates a count starting from -100 with a step of 50","status":"passed","title":"creates a count starting from -100 with a step of 50","duration":8,"failureMessages":[]},{"ancestorTitles":[""],"fullName":" creates a count starting from 100 with a step of -50","status":"passed","title":"creates a count starting from 100 with a step of -50","duration":2,"failureMessages":[]},{"ancestorTitles":[""],"fullName":" creates a count starting from 141592653589793 with a step of 141592653589793","status":"passed","title":"creates a count starting from 141592653589793 with a step of 141592653589793","duration":1,"failureMessages":[]},{"ancestorTitles":[""],"fullName":" creates a count starting from 0 with a step of 0","status":"passed","title":"creates a count starting from 0 with a step of 0","duration":1,"failureMessages":[]},{"ancestorTitles":[""],"fullName":" creates a count starting from 5 with a step of 0","status":"passed","title":"creates a count starting from 5 with a step of 0","duration":1,"failureMessages":[]},{"ancestorTitles":[""],"fullName":" when no first argument passed, defaults to start a count starting from 0 with a step of 5","status":"passed","title":"when no first argument passed, defaults to start a count starting from 0 with a step of 5","duration":1,"failureMessages":[]},{"ancestorTitles":[""],"fullName":" when no second argument passed, defaults to start a count starting from 5 with a step of 1","status":"passed","title":"when no second argument passed, defaults to start a count starting from 5 with a step of 1","duration":1,"failureMessages":[]}],"startTime":1723061932651,"endTime":1723061932666,"status":"passed","message":"","name":"/Users/felipesilva/Flip/centered_counter/src/proposed-factory.test.ts"}]}
